Split EddystoneService.h into multiple files
